[Decomp][Spell] Cohesion split: extract SpellAuraShapeshift handlers to SpellAuraShapeshift.cpp
