+++ src/game/WorldHandlers/SpellAuras.cpp
@@ -2131,880 +2131,21 @@ void Aura::TriggerSpellWithValue()
 /*********************************************************/
 
 
-/**
- * @brief Applies or removes a mounted display from the target.
- *
- * @param apply True to mount; false to unmount.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraMounted(bool apply, bool Real)
-{
-    // only at real add/remove aura
-    if (!Real)
-    {
-        return;
-    }
 
-    Unit* target = GetTarget();
 
-    if (apply)
-    {
-        CreatureInfo const* ci = ObjectMgr::GetCreatureTemplate(m_modifier.m_miscvalue);
-        if (!ci)
-        {
-            sLog.outErrorDb("AuraMounted: `creature_template`='%u' not found in database (only need it modelid)", m_modifier.m_miscvalue);
-            return;
-        }
 
-        uint32 display_id = Creature::ChooseDisplayId(ci);
-        CreatureModelInfo const* minfo = sObjectMgr.GetCreatureModelRandomGender(display_id);
-        if (minfo)
-        {
-            display_id = minfo->modelid;
-        }
 
-        target->Mount(display_id, GetId());
-    }
-    else
-    {
-        target->Unmount(true);
-    }
-}
 
-/**
- * @brief Applies or removes water walking on the target.
- *
- * @param apply True to enable water walking; false to disable it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraWaterWalk(bool apply, bool Real)
-{
-    // only at real add/remove aura
-    if (!Real)
-    {
-        return;
-    }
 
-    GetTarget()->SetWaterWalk(apply);
-}
 
-/**
- * @brief Applies or removes feather fall on the target.
- *
- * @param apply True to enable feather fall; false to disable it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraFeatherFall(bool apply, bool Real)
-{
-    // only at real add/remove aura
-    if (!Real)
-    {
-        return;
-    }
 
-    GetTarget()->SetFeatherFall(apply);
-}
 
-/**
- * @brief Applies or removes hovering movement state.
- *
- * @param apply True to enable hover; false to disable it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraHover(bool apply, bool Real)
-{
-    // only at real add/remove aura
-    if (!Real)
-    {
-        return;
-    }
 
-    GetTarget()->SetHover(apply);
-}
 
-/**
- * @brief Refreshes client breathing timers for the target.
- */
-void Aura::HandleWaterBreathing(bool /*apply*/, bool /*Real*/)
-{
-    // update timers in client
-    if (GetTarget()->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)GetTarget())->UpdateMirrorTimers();
-    }
-}
 
-/**
- * @brief Applies or removes a shapeshift form and its related state changes.
- *
- * @param apply True to enter the form; false to leave it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraModShapeshift(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
 
-    ShapeshiftForm form = ShapeshiftForm(m_modifier.m_miscvalue);
 
-    SpellShapeshiftFormEntry const* ssEntry = sSpellShapeshiftFormStore.LookupEntry(form);
-    if (!ssEntry)
-    {
-        sLog.outError("Unknown shapeshift form %u in spell %u", form, GetId());
-        return;
-    }
 
-    uint32 modelid = 0;
-    Powers PowerType = POWER_MANA;
-    Unit* target = GetTarget();
-
-    // remove SPELL_AURA_EMPATHY
-    target->RemoveSpellsCausingAura(SPELL_AURA_EMPATHY);
-
-    if (ssEntry->modelID_A)
-    {
-        // i will asume that creatures will always take the defined model from the dbc
-        // since no field in creature_templates describes wether an alliance or
-        // horde modelid should be used at shapeshifting
-        if (target->GetTypeId() != TYPEID_PLAYER)
-        {
-            modelid = ssEntry->modelID_A;
-        }
-        else
-        {
-            // players are a bit different since the dbc has seldomly an horde modelid
-            if (Player::TeamForRace(target->getRace()) == HORDE)
-            {
-                // get model for race ( in 2.2.4 no horde models in dbc field, only 0 in it
-                modelid = sObjectMgr.GetModelForRace(ssEntry->modelID_A, target->getRaceMask());
-            }
-
-            // nothing found in above, so use default
-            if (!modelid)
-            {
-                modelid = ssEntry->modelID_A;
-            }
-        }
-    }
-
-    // remove polymorph before changing display id to keep new display id
-    switch (form)
-    {
-        case FORM_CAT:
-        case FORM_TREE:
-        case FORM_TRAVEL:
-        case FORM_AQUA:
-        case FORM_BEAR:
-        case FORM_DIREBEAR:
-        case FORM_FLIGHT_EPIC:
-        case FORM_FLIGHT:
-        case FORM_MOONKIN:
-        {
-            // remove movement affects
-            target->RemoveSpellsCausingAura(SPELL_AURA_MOD_ROOT, GetHolder());
-            Unit::AuraList const& slowingAuras = target->GetAurasByType(SPELL_AURA_MOD_DECREASE_SPEED);
-            for (Unit::AuraList::const_iterator iter = slowingAuras.begin(); iter != slowingAuras.end();)
-            {
-                SpellEntry const* aurSpellInfo = (*iter)->GetSpellProto();
-
-                uint32 aurMechMask = GetAllSpellMechanicMask(aurSpellInfo);
-
-                // If spell that caused this aura has Croud Control or Daze effect
-                if ((aurMechMask & MECHANIC_NOT_REMOVED_BY_SHAPESHIFT) ||
-                    // some Daze spells have these parameters instead of MECHANIC_DAZE (skip snare spells)
-                    (aurSpellInfo->SpellIconID == 15 && aurSpellInfo->Dispel == 0 &&
-                     (aurMechMask & (1 << (MECHANIC_SNARE - 1))) == 0))
-                {
-                    ++iter;
-                    continue;
-                }
-
-                // All OK, remove aura now
-                target->RemoveAurasDueToSpellByCancel(aurSpellInfo->Id);
-                iter = slowingAuras.begin();
-            }
-
-            // and polymorphic affects
-            if (target->IsPolymorphed())
-            {
-                target->RemoveAurasDueToSpell(target->GetTransform());
-            }
-
-            //no break here
-        }
-        case FORM_GHOSTWOLF:
-        {
-            // remove water walk aura. TODO:: there is probably better way to do this
-            target->RemoveSpellsCausingAura(SPELL_AURA_WATER_WALK);
-
-            break;
-        }
-        default:
-            break;
-    }
-
-    if (apply)
-    {
-        // remove other shapeshift before applying a new one
-        target->RemoveSpellsCausingAura(SPELL_AURA_MOD_SHAPESHIFT, GetHolder());
-
-        if (modelid > 0)
-        {
-            target->SetObjectScale(DEFAULT_OBJECT_SCALE);
-            target->SetDisplayId(modelid);
-        }
-
-        // now only powertype must be set
-        switch (form)
-        {
-            case FORM_CAT:
-                PowerType = POWER_ENERGY;
-                break;
-            case FORM_BEAR:
-            case FORM_DIREBEAR:
-            case FORM_BATTLESTANCE:
-            case FORM_BERSERKERSTANCE:
-            case FORM_DEFENSIVESTANCE:
-                PowerType = POWER_RAGE;
-                break;
-            default:
-                break;
-        }
-
-        if (PowerType != POWER_MANA)
-        {
-            // reset power to default values only at power change
-            if (target->GetPowerType() != PowerType)
-            {
-                target->SetPowerType(PowerType);
-            }
-
-            switch (form)
-            {
-                case FORM_CAT:
-                case FORM_BEAR:
-                case FORM_DIREBEAR:
-                {
-                    // get furor proc chance
-                    int32 furorChance = 0;
-                    Unit::AuraList const& mDummy = target->GetAurasByType(SPELL_AURA_DUMMY);
-                    for (Unit::AuraList::const_iterator i = mDummy.begin(); i != mDummy.end(); ++i)
-                    {
-                        if ((*i)->GetSpellProto()->SpellIconID == 238)
-                        {
-                            furorChance = (*i)->GetModifier()->m_amount;
-                            break;
-                        }
-                    }
-
-                    if (m_modifier.m_miscvalue == FORM_CAT)
-                    {
-                        target->SetPower(POWER_ENERGY, 0);
-                        if (irand(1, 100) <= furorChance)
-                        {
-                            target->CastSpell(target, 17099, true, NULL, this);
-                        }
-                    }
-                    else
-                    {
-                        target->SetPower(POWER_RAGE, 0);
-                        if (irand(1, 100) <= furorChance)
-                        {
-                            target->CastSpell(target, 17057, true, NULL, this);
-                        }
-                    }
-                    break;
-                }
-                case FORM_BATTLESTANCE:
-                case FORM_DEFENSIVESTANCE:
-                case FORM_BERSERKERSTANCE:
-                {
-                    uint32 Rage_val = 0;
-                    // Stance mastery + Tactical mastery (both passive, and last have aura only in defense stance, but need apply at any stance switch)
-                    if (target->GetTypeId() == TYPEID_PLAYER)
-                    {
-                        PlayerSpellMap const& sp_list = ((Player*)target)->GetSpellMap();
-                        for (PlayerSpellMap::const_iterator itr = sp_list.begin(); itr != sp_list.end(); ++itr)
-                        {
-                            if (itr->second.state == PLAYERSPELL_REMOVED)
-                            {
-                                continue;
-                            }
-
-                            SpellEntry const* spellInfo = sSpellStore.LookupEntry(itr->first);
-                            if (spellInfo && spellInfo->SpellFamilyName == SPELLFAMILY_WARRIOR && spellInfo->SpellIconID == 139)
-                            {
-                                Rage_val += target->CalculateSpellDamage(target, spellInfo, EFFECT_INDEX_0) * 10;
-                            }
-                        }
-                    }
-
-                    if (target->GetPower(POWER_RAGE) > Rage_val)
-                    {
-                        target->SetPower(POWER_RAGE, Rage_val);
-                    }
-                    break;
-                }
-                default:
-                    break;
-            }
-        }
-
-        target->SetShapeshiftForm(form);
-
-        // a form can give the player a new castbar with some spells.. this is a clientside process..
-        // serverside just needs to register the new spells so that player isn't kicked as cheater
-        if (target->GetTypeId() == TYPEID_PLAYER)
-            for (uint32 i = 0; i < 8; ++i)
-            {
-                if (ssEntry->spellId[i])
-                {
-                    ((Player*)target)->addSpell(ssEntry->spellId[i], true, false, false, false);
-                }
-            }
-    }
-    else
-    {
-        if (modelid > 0)
-        {
-            target->SetDisplayId(target->GetNativeDisplayId());
-        }
-
-        if (target->getClass() == CLASS_DRUID)
-        {
-            target->SetPowerType(POWER_MANA);
-        }
-
-        target->SetShapeshiftForm(FORM_NONE);
-
-        switch (form)
-        {
-                // Nordrassil Harness - bonus
-            case FORM_BEAR:
-            case FORM_DIREBEAR:
-            case FORM_CAT:
-                if (Aura* dummy = target->GetDummyAura(37315))
-                {
-                    target->CastSpell(target, 37316, true, NULL, dummy);
-                }
-                break;
-                // Nordrassil Regalia - bonus
-            case FORM_MOONKIN:
-                if (Aura* dummy = target->GetDummyAura(37324))
-                {
-                    target->CastSpell(target, 37325, true, NULL, dummy);
-                }
-                break;
-            default:
-                break;
-        }
-
-        // look at the comment in apply-part
-        if (target->GetTypeId() == TYPEID_PLAYER)
-            for (uint32 i = 0; i < 8; ++i)
-            {
-                if (ssEntry->spellId[i])
-                {
-                    ((Player*)target)->removeSpell(ssEntry->spellId[i], false, false, false);
-                }
-            }
-    }
-
-    // adding/removing linked auras
-    // add/remove the shapeshift aura's boosts
-    HandleShapeshiftBoosts(apply);
-
-    if (target->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)target)->InitDataForForm();
-    }
-}
-
-/**
- * @brief Applies or removes a transform model effect.
- *
- * @param apply True to apply the transform; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraTransform(bool apply, bool Real)
-{
-    Unit* target = GetTarget();
-    if (apply)
-    {
-        // special case (spell specific functionality)
-        if (m_modifier.m_miscvalue == 0)
-        {
-            switch (GetId())
-            {
-                case 16739:                                 // Orb of Deception
-                {
-                    uint32 orb_model = target->GetNativeDisplayId();
-                    switch (orb_model)
-                    {
-                            // Troll Female
-                        case 1479: target->SetDisplayId(10134); break;
-                            // Troll Male
-                        case 1478: target->SetDisplayId(10135); break;
-                            // Tauren Male
-                        case 59:   target->SetDisplayId(10136); break;
-                            // Human Male
-                        case 49:   target->SetDisplayId(10137); break;
-                            // Human Female
-                        case 50:   target->SetDisplayId(10138); break;
-                            // Orc Male
-                        case 51:   target->SetDisplayId(10139); break;
-                            // Orc Female
-                        case 52:   target->SetDisplayId(10140); break;
-                            // Dwarf Male
-                        case 53:   target->SetDisplayId(10141); break;
-                            // Dwarf Female
-                        case 54:   target->SetDisplayId(10142); break;
-                            // NightElf Male
-                        case 55:   target->SetDisplayId(10143); break;
-                            // NightElf Female
-                        case 56:   target->SetDisplayId(10144); break;
-                            // Undead Female
-                        case 58:   target->SetDisplayId(10145); break;
-                            // Undead Male
-                        case 57:   target->SetDisplayId(10146); break;
-                            // Tauren Female
-                        case 60:   target->SetDisplayId(10147); break;
-                            // Gnome Male
-                        case 1563: target->SetDisplayId(10148); break;
-                            // Gnome Female
-                        case 1564: target->SetDisplayId(10149); break;
-                            // BloodElf Female
-                        case 15475: target->SetDisplayId(17830); break;
-                            // BloodElf Male
-                        case 15476: target->SetDisplayId(17829); break;
-                            // Dranei Female
-                        case 16126: target->SetDisplayId(17828); break;
-                            // Dranei Male
-                        case 16125: target->SetDisplayId(17827); break;
-                        default: break;
-                    }
-                    break;
-                }
-                case 42365:                                 // Murloc costume
-                    target->SetDisplayId(21723);
-                    break;
-                    // case 44186:                          // Gossip NPC Appearance - All, Brewfest
-                    // break;
-                    // case 48305:                          // Gossip NPC Appearance - All, Spirit of Competition
-                    // break;
-                case 50517:                                 // Dread Corsair
-                case 51926:                                 // Corsair Costume
-                {
-                    // expected for players
-                    uint32 race = target->getRace();
-
-                    switch (race)
-                    {
-                        case RACE_HUMAN:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25037 : 25048);
-                            break;
-                        case RACE_ORC:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25039 : 25050);
-                            break;
-                        case RACE_DWARF:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25034 : 25045);
-                            break;
-                        case RACE_NIGHTELF:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25038 : 25049);
-                            break;
-                        case RACE_UNDEAD:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25042 : 25053);
-                            break;
-                        case RACE_TAUREN:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25040 : 25051);
-                            break;
-                        case RACE_GNOME:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25035 : 25046);
-                            break;
-                        case RACE_TROLL:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25041 : 25052);
-                            break;
-                        case RACE_GOBLIN:                   // not really player race (3.x), but model exist
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25036 : 25047);
-                            break;
-                        case RACE_BLOODELF:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25032 : 25043);
-                            break;
-                        case RACE_DRAENEI:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25033 : 25044);
-                            break;
-                    }
-
-                    break;
-                }
-                // case 50531:                              // Gossip NPC Appearance - All, Pirate Day
-                // break;
-                // case 51010:                              // Dire Brew
-                // break;
-                default:
-                    sLog.outError("Aura::HandleAuraTransform, spell %u does not have creature entry defined, need custom defined model.", GetId());
-                    break;
-            }
-        }
-        else                                                // m_modifier.m_miscvalue != 0
-        {
-            uint32 model_id;
-
-            CreatureInfo const* ci = ObjectMgr::GetCreatureTemplate(m_modifier.m_miscvalue);
-            if (!ci)
-            {
-                model_id = 16358;                           // pig pink ^_^
-                sLog.outError("Auras: unknown creature id = %d (only need its modelid) Form Spell Aura Transform in Spell ID = %d", m_modifier.m_miscvalue, GetId());
-            }
-            else
-            {
-                model_id = Creature::ChooseDisplayId(ci);    // Will use the default model here
-            }
-
-            target->SetDisplayId(model_id);
-
-            // creature case, need to update equipment if additional provided
-            if (ci && target->GetTypeId() == TYPEID_UNIT)
-            {
-                ((Creature*)target)->LoadEquipment(ci->EquipmentTemplateId, false);
-            }
-
-            // Dragonmaw Illusion (set mount model also)
-            if (GetId() == 42016 && target->GetMountID() && !target->GetAurasByType(SPELL_AURA_MOD_FLIGHT_SPEED_MOUNTED).empty())
-            {
-                target->SetUInt32Value(UNIT_FIELD_MOUNTDISPLAYID, 16314);
-            }
-        }
-
-        // update active transform spell only not set or not overwriting negative by positive case
-        if (!target->GetTransform() || !IsPositiveSpell(GetId()) || IsPositiveSpell(target->GetTransform()))
-        {
-            target->SetTransform(GetId());
-        }
-
-        // polymorph case
-        if (Real && target->GetTypeId() == TYPEID_PLAYER && target->IsPolymorphed())
-        {
-            // for players, start regeneration after 1s (in polymorph fast regeneration case)
-            // only if caster is Player (after patch 2.4.2)
-            if (GetCasterGuid().IsPlayer())
-            {
-                ((Player*)target)->setRegenTimer(1 * IN_MILLISECONDS);
-            }
-
-            // dismount polymorphed target (after patch 2.4.2)
-            if (target->IsMounted())
-            {
-                target->RemoveSpellsCausingAura(SPELL_AURA_MOUNTED, GetHolder());
-            }
-        }
-    }
-    else                                                    // !apply
-    {
-        // ApplyModifier(true) will reapply it if need
-        target->SetTransform(0);
-        target->SetDisplayId(target->GetNativeDisplayId());
-
-        // apply default equipment for creature case
-        if (target->GetTypeId() == TYPEID_UNIT)
-        {
-            ((Creature*)target)->LoadEquipment(((Creature*)target)->GetCreatureInfo()->EquipmentTemplateId, true);
-        }
-
-        // re-apply some from still active with preference negative cases
-        Unit::AuraList const& otherTransforms = target->GetAurasByType(SPELL_AURA_TRANSFORM);
-        if (!otherTransforms.empty())
-        {
-            // look for other transform auras
-            Aura* handledAura = *otherTransforms.begin();
-            for (Unit::AuraList::const_iterator i = otherTransforms.begin(); i != otherTransforms.end(); ++i)
-            {
-                // negative auras are preferred
-                if (!IsPositiveSpell((*i)->GetSpellProto()->Id))
-                {
-                    handledAura = *i;
-                    break;
-                }
-            }
-            handledAura->ApplyModifier(true);
-        }
-
-        // Dragonmaw Illusion (restore mount model)
-        if (GetId() == 42016 && target->GetMountID() == 16314)
-        {
-            if (!target->GetAurasByType(SPELL_AURA_MOUNTED).empty())
-            {
-                uint32 cr_id = target->GetAurasByType(SPELL_AURA_MOUNTED).front()->GetModifier()->m_miscvalue;
-                if (CreatureInfo const* ci = ObjectMgr::GetCreatureTemplate(cr_id))
-                {
-                    uint32 display_id = Creature::ChooseDisplayId(ci);
-                    CreatureModelInfo const* minfo = sObjectMgr.GetCreatureModelRandomGender(display_id);
-                    if (minfo)
-                    {
-                        display_id = minfo->modelid;
-                    }
-
-                    target->SetUInt32Value(UNIT_FIELD_MOUNTDISPLAYID, display_id);
-                }
-            }
-        }
-    }
-}
-
-/**
- * @brief Applies or removes a forced reputation reaction for a player.
- *
- * @param apply True to apply the forced reaction; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleForceReaction(bool apply, bool Real)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (!Real)
-    {
-        return;
-    }
-
-    Player* player = (Player*)GetTarget();
-
-    uint32 faction_id = m_modifier.m_miscvalue;
-    ReputationRank faction_rank = ReputationRank(m_modifier.m_amount);
-
-    player->GetReputationMgr().ApplyForceReaction(faction_id, faction_rank, apply);
-    player->GetReputationMgr().SendForceReactions();
-
-    // stop fighting if at apply forced rank friendly or at remove real rank friendly
-    if ((apply && faction_rank >= REP_FRIENDLY) || (!apply && player->GetReputationRank(faction_id) >= REP_FRIENDLY))
-    {
-        player->StopAttackFaction(faction_id);
-    }
-}
-
-/**
- * @brief Applies or removes a player skill bonus from the aura.
- *
- * @param apply True to apply the bonus; false to remove it.
- */
-void Aura::HandleAuraModSkill(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    uint32 prot = GetSpellProto()->EffectMiscValue[m_effIndex];
-    int32 points = GetModifier()->m_amount;
-
-    ((Player*)GetTarget())->ModifySkillBonus(prot, (apply ? points : -points), m_modifier.m_auraname == SPELL_AURA_MOD_SKILL_TALENT);
-    if (prot == SKILL_DEFENSE)
-    {
-        ((Player*)GetTarget())->UpdateDefenseBonusesMod();
-    }
-}
-
-/**
- * @brief Awards the configured item when a channel-death aura ends by death.
- *
- * @param apply True on application; false on removal.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleChannelDeathItem(bool apply, bool Real)
-{
-    if (Real && !apply)
-    {
-        if (m_removeMode != AURA_REMOVE_BY_DEATH)
-        {
-            return;
-        }
-        // Item amount
-        if (m_modifier.m_amount <= 0)
-        {
-            return;
-        }
-
-        SpellEntry const* spellInfo = GetSpellProto();
-        if (spellInfo->EffectItemType[m_effIndex] == 0)
-        {
-            return;
-        }
-
-        Unit* victim = GetTarget();
-        Unit* caster = GetCaster();
-        if (!caster || caster->GetTypeId() != TYPEID_PLAYER)
-        {
-            return;
-        }
-
-        // Soul Shard (target req.)
-        if (spellInfo->EffectItemType[m_effIndex] == 6265)
-        {
-            // Only from non-grey units
-            if (!((Player*)caster)->isHonorOrXPTarget(victim) ||
-                (victim->GetTypeId() == TYPEID_UNIT && !((Creature*)victim)->IsTappedBy((Player*)caster)))
-            {
-                return;
-            }
-        }
-
-        // Adding items
-        uint32 noSpaceForCount = 0;
-        uint32 count = m_modifier.m_amount;
-
-        ItemPosCountVec dest;
-        InventoryResult msg = ((Player*)caster)->CanStoreNewItem(NULL_BAG, NULL_SLOT, dest, spellInfo->EffectItemType[m_effIndex], count, &noSpaceForCount);
-        if (msg != EQUIP_ERR_OK)
-        {
-            count -= noSpaceForCount;
-            ((Player*)caster)->SendEquipError(msg, NULL, NULL, spellInfo->EffectItemType[m_effIndex]);
-            if (count == 0)
-            {
-                return;
-            }
-        }
-
-        Item* newitem = ((Player*)caster)->StoreNewItem(dest, spellInfo->EffectItemType[m_effIndex], true);
-        ((Player*)caster)->SendNewItem(newitem, count, true, true);
-    }
-}
-
-/**
- * @brief Redirects the caster camera to the target while the aura is active.
- *
- * @param apply True to bind sight; false to restore normal view.
- */
-void Aura::HandleBindSight(bool apply, bool /*Real*/)
-{
-    Unit* caster = GetCaster();
-    if (!caster || caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Camera& camera = ((Player*)caster)->GetCamera();
-    if (apply)
-    {
-        camera.SetView(GetTarget());
-    }
-    else
-    {
-        camera.ResetView();
-    }
-}
-
-/**
- * @brief Redirects the caster camera for farsight while the aura is active.
- *
- * @param apply True to enable farsight; false to restore normal view.
- */
-void Aura::HandleFarSight(bool apply, bool /*Real*/)
-{
-    Unit* caster = GetCaster();
-    if (!caster || caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Camera& camera = ((Player*)caster)->GetCamera();
-    if (apply)
-    {
-        camera.SetView(GetTarget());
-    }
-    else
-    {
-        camera.ResetView();
-    }
-}
-
-/**
- * @brief Applies or removes creature tracking flags on a player.
- *
- * @param apply True to enable tracking; false to disable it.
- */
-void Aura::HandleAuraTrackCreatures(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        GetTarget()->RemoveNoStackAurasDueToAuraHolder(GetHolder());
-    }
-
-    if (apply)
-    {
-        GetTarget()->SetFlag(PLAYER_TRACK_CREATURES, uint32(1) << (m_modifier.m_miscvalue - 1));
-    }
-    else
-    {
-        GetTarget()->RemoveFlag(PLAYER_TRACK_CREATURES, uint32(1) << (m_modifier.m_miscvalue - 1));
-    }
-}
-
-/**
- * @brief Applies or removes resource tracking flags on a player.
- *
- * @param apply True to enable tracking; false to disable it.
- */
-void Aura::HandleAuraTrackResources(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        GetTarget()->RemoveNoStackAurasDueToAuraHolder(GetHolder());
-    }
-
-    if (apply)
-    {
-        GetTarget()->SetFlag(PLAYER_TRACK_RESOURCES, uint32(1) << (m_modifier.m_miscvalue - 1));
-    }
-    else
-    {
-        GetTarget()->RemoveFlag(PLAYER_TRACK_RESOURCES, uint32(1) << (m_modifier.m_miscvalue - 1));
-    }
-}
-
-/**
- * @brief Applies or removes stealthed-unit tracking on a player.
- *
- * @param apply True to enable tracking; false to disable it.
- */
-void Aura::HandleAuraTrackStealthed(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        GetTarget()->RemoveNoStackAurasDueToAuraHolder(GetHolder());
-    }
-
-    GetTarget()->ApplyModByteFlag(PLAYER_FIELD_BYTES, 0, PLAYER_FIELD_BYTE_TRACK_STEALTHED, apply);
-}
-
-/**
- * @brief Applies or removes a scale modifier and refreshes model data.
- *
- * @param apply True to apply the scale change; false to remove it.
- */
-void Aura::HandleAuraModScale(bool apply, bool /*Real*/)
-{
-    GetTarget()->ApplyPercentModFloatValue(OBJECT_FIELD_SCALE_X, float(m_modifier.m_amount), apply);
-    GetTarget()->UpdateModelData();
-}
